Fixes #1: filter_genomic_location_by_chr_name()
There was a bug in the function filter_genomic_location_by_chr_name()
that propagated to higher-level functions that deal with snp queries.
The bug stemmed from the assumption that in the locations dataframe
the column _links.snps.href was always created. However, in some cases
the column _links.snps is created instead. This is much less often
but happens because one snp might have more than one link to other
snps. Example: rs35252396 and rs6470588 have both listed each other
in _links.snps. Usually only an href is present, to itself.